Clarified location of hooks
`code-rack` uses `dirname $0` to determine where the hook scripts are, not /etc/dehydrated/hooks as the documentation says.

I have changed the wording to make clear that it looks where ever `code-rack` is, instead.  Alternatively, it could be rewritten to make clear that **code-rack** must be installed in /etc/dehydrated/hooks for the instructions to work.